[TASK] Behavioral, package and testbench for the busMaster
The aim of the behavorial design is to stimulate bus slaves via a
.txt file. So the busMaster has to parse and interpret the commands and
set the outputs on the rising clock edge.
The package offers many function and results in a more structured design.
It is easily possible to add further instruction in the stimuli file
For additional information refer to busMasterPkg.

Change-Id: I7c7ffb6aef6f7254b7b0126b314f41f1202a2b3d
Reviewed-on: http://brateggevoat1/gerrit/1096
Tested-by: Jenkins <jenkins@brateggevoat1.br-automation.co.at>
Reviewed-by: Joerg Zelenka <joerg.zelenka@br-automation.com>